Update ReadMe.txt
Traffic Road Accidents cause several severe injuries or even deaths every year among the United States of America. Deaths and injury damage can be reduced if the injured patient is treated in the least time possible. However, Emergency Medical Systems are limited and information on patient condition at the time of accident notification is usually not available. Thus, miss-interpretation of severity will lead to death or injure severity. The proposed research shows a working algorithm for Real-Time classification of severe accidents using little easy automatically collectible information such as Temperature and Humidity. Research methodology is explained for transparency and reproducibility of results as well. The final model was built with K-Nearest Neighbors (KNN) showing high precision and recall scores and could be deployed for States such as California, Texas, and Florida. The developed model can be combined with on-car severity prediction technologies or used alone if not available. Model extension to other areas as well as improvement of already achieved results for previously mentioned states can easily be performed with KNN. Extension to further countries is possible although the country development state should be taken into consideration as it affects the model.